Split modify logic up into extend/release/common files
