+++ src/mca/ras/slurm/ras_slurm.h
@@ -42,12 +42,21 @@
 
 BEGIN_C_DECLS
 
+/* Features requiring JSON parser*/
 int prte_ras_slurm_extract_job_fields(pmix_hash_table_t *values_table);
 int prte_ras_slurm_add_modified_resources(const char *slurm_jobid, pmix_list_t *node_list);
 int prte_ras_slurm_wait_resources(const char *slurm_jobid);
-int prte_ras_slurm_launch_expander_job(pmix_hash_table_t *fields);
-int prte_ras_slurm_assign_new_session(const char *slurm_jobid, const char *alloc_refid, pmix_list_t *node_list);
-int prte_ras_slurm_reject_node_duplicates(pmix_list_t *node_list);
+
+/* Features to serve request extension */
+int prte_ras_slurm_serve_extend_req(prte_pmix_server_req_t *req);
+
+/* Features to serve request release */
+int prte_ras_slurm_serve_release_req(prte_pmix_server_req_t *req);
+
+/* Common modify extend/release features */
+int prte_ras_slurm_validate_jobid(const char *slurm_jobid);
+int prte_ras_slurm_convert_jobid(const char *slurm_jobid, uint32_t *slurm_jobid_numeric);
+int prte_ras_slurm_kill_job(const char *slurm_jobid, char *err_msg);
 
 typedef struct {
     prte_ras_base_component_t super;
